Add 3 functions to make game playable
Three functions were added, as well as two helper functions. The first
function added was computerPlay, which made a random number from 0-2,
and choosing either rock, paper, or scissors based off of that number.
The second function was playRound, which simulated a round of the game.
A prompt is asked for the user to input their choice (it's not case-
sensitive.) Depending on the player's choice, and the computer's
choice, a string is returned with a message that you either won or lost.
The third function, game() simulates a game with 5 rounds. It tells
you if you won or lost based off the final score. There are two
helper functions which update the computer and player's score after
each round. This is what is used to figure out who won.